Prepare diamond implementation pattern for audit (#2867)
This PR updates our diamond implementation to a modular deployment
pattern by removing the default external facets within Diamond and
allowing developers to add their own baseFacets at deployment time.

It also updates the diamond file structure to be composed of 3 main
files, a storage file, a base file and a facet file.

**Storage**
This file holds the storage declaration and the `.layout()` function
that allows a base contract to access the storage slots

**Base**
This file holds the majority of the internal business logic and talks
directly to our **Storage File**

**Facet**
This is the external facing implementation, you can inherit multiple
Base files into this Facet in order to modular functionality. For
example, you can have a `ERC721Facet` that inherits `ERC721Base.sol`,
and `OwnableBase` so you can have modifiers like `onlyOwner` within your
`mint` function

+++ packages/contracts/test/diamond/initializable/Initializable.t.sol
@@ -18,7 +18,7 @@ contract Mock is Initializable {
 
   function reinit() external reinitializer(2) {}
 
-  function getVersion() external view returns (uint8) {
+  function getVersion() external view returns (uint32) {
     return _getInitializedVersion();
   }
 }
